Populate some new CalcInfos fields on Titanium/USB & TI84+/USB.
git-svn-id: http://svn.tilp.info/repos/tilp@3994

+++ libticalcs/trunk/src/calc_nsp.c
@@ -508,7 +508,7 @@ static int		get_version	(CalcHandle* handle, CalcInfos* infos)
 	infos->mask |= INFOS_HW_VERSION;
 
 	i = 52;
-	infos->run_level = GUINT16_FROM_BE(*((uint16_t *)(data + i)));
+	infos->run_level = (uint8_t)GUINT16_FROM_BE(*((uint16_t *)(data + i)));
 	infos->mask |= INFOS_RUN_LEVEL;
 
 	i = 58;
